+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Version 1.139 (not yet released)
 
 * Added Action: Relationship: Create with one or more contacts
 * Added validators for Is Empty and Parameter Has Value.
+* Added Action: Send email to a group.
 
 Version 1.138
 -------------